requirements.txt
HTMLTestRunner-rv==1.1.2
matplotlib==3.8.2
pandas==2.1.4